getLoginFailureCount() 메서드의 return type int -> short로

+++ member-service/src/main/java/com/example/member/dto/Member.java
@@ -50,7 +50,7 @@ public class Member {
 	public void setPhone(String phone) {
 		this.phone = phone;
 	}
-	public int getLoginFailureCount() {
+	public short getLoginFailureCount() {
 		return loginFailureCount;
 	}
 	public void setLoginFailureCount(int loginFailureCount) {
